Add Prometheus to realize the function of monitoring and alarming

+++ deployment/kubernetes/alerting/stop_alerting.sh
@@ -2,12 +2,11 @@
 
 DIR=$(dirname $(readlink -f "$0"))
 
-# Try command  for test command result.
-
 # Set Bash color
 ECHO_PREFIX_INFO="\033[1;32;40mINFO...\033[0;0m"
 ECHO_PREFIX_ERROR="\033[1;31;40mError...\033[0;0m"
 
+# Try command for test command result
 function try_command {
     "$@" 2> /dev/null
     status=$?
@@ -26,9 +25,14 @@ fi
 
 set +e
 try_command hash kubectl > /dev/null
-for i in $(ls $DIR/*.yaml); do
-    kubectl delete -f "$i" &> /dev/null
-done
 set -e
 
-rm -rf $DIR/*[0-9].yaml
+if (kubectl get namespace | awk '{print $1}' | grep -q "kube-prometheus"); then
+    kubectl delete -f "$DIR/namespace/namespace.yaml"
+
+    for i in $(find "$DIR" -path "$DIR/namespace" -a -prune -o -name "*.yaml" -print); do
+        kubectl delete -f "$i" &> /dev/null
+    done
+fi
+
+echo "Prometheus are stopping..."
